+++ src/edge.go
@@ -8,11 +8,11 @@ func (e *Edge) To() interface{} {
 	return e.to.Data()
 }
 
-func (e *Edge) Cost() int64 {
+func (e *Edge) Cost() float64 {
 	return e.cost
 }
 
-func (e *Edge) SetCost(c int64) bool {
+func (e *Edge) SetCost(c float64) bool {
 	if e == nil {
 		return false
 	}
